fix: make pasteurbot follow conventional commits (#8)
#### Relevant issue or PR
<!-- If the changes resolve an issue or follow some other PR, link to
them here. GitHub references are strongly preferred, but links to other
resources, such as Jira or Confluence, can also be used. Only link
something if it is directly relevant. -->

Closes #7.

#### Description of changes
<!-- Add a high-level description of changes, focusing on the *what* and
*why*. -->

Changed PR titles and commit messages by PasteurBot in workflows with
actions `peter-evans/create-pull-request`. Prepended with `"chore: "`,
and made all lower-case.

#### Testing done
<!-- Describe how the changes were tested; e.g., "CI passes", "Tested
manually in stagingrepo#123", screenshots of a terminal session that
verify the changes, or any other evidence of testing the changes. -->

n/a