fix: support json pretty-print for single elem array outputs (#15)
#### Relevant issue or PR
<!-- If the changes resolve an issue or follow some other PR, link to
them here. GitHub references are strongly preferred, but links to other
resources, such as Jira or Confluence, can also be used. Only link
something if it is directly relevant. -->
Cherry picked from #9.

#### Description of changes
<!-- Add a high-level description of changes, focusing on the *what* and
*why*. -->
Fixes issue with JSON pretty-printing of Tesseract outputs containing
zero-dimensional arrays. Prior to pretty printing, the leaves of the
Tesseract outputs are visited, and zero-dimensional arrays are replaced
with scalars.

I've just noticed that this actually happens in-place, which I would
prefer not to do.

To do:

- [x] Deep copy the output dictionary before pretty-printing
- [x] Add e2e automated test for a Tesseract which outputs a 0-dim array

#### Testing done
<!-- Describe how the changes were tested; e.g., "CI passes", "Tested
manually in stagingrepo#123", screenshots of a terminal session that
verify the changes, or any other evidence of testing the changes. -->

Branched from `main` to add an e2e test prior to the fix. Running
`pytest` confirmed that the test failed. Once it was merged into
`jacan/zero-dim-array-json`, running `pytest` resulted in all tests
passing.