fix: use trame backend for stpyvista (#18)
#### Relevant issue or PR
<!-- If the changes resolve an issue or follow some other PR, link to
them here. GitHub references are strongly preferred, but links to other
resources, such as Jira or Confluence, can also be used. Only link
something if it is directly relevant. -->
Replaces #14.

#### Description of changes
<!-- Add a high-level description of changes, focusing on the *what* and
*why*. -->
Switches the backend for PyVista to trame instead of panel. This was
necessary as panel appears to render the visualisations zoomed in, and
would break depending on which form of interaction (panning, zooming,
rotating) first occurred on the visualisation. The trame backend does
not have these issues.

#### Testing done
<!-- Describe how the changes were tested; e.g., "CI passes", "Tested
manually in stagingrepo#123", screenshots of a terminal session that
verify the changes, or any other evidence of testing the changes. -->

Manual testing with web UI.