test: add stpyvista trame backend test with macos gha (#19)
#### Description of changes
<!-- Add a high-level description of changes, focusing on the *what* and
*why*. -->
The `trame` backend to PyVista causes `stpyvista` to fail on macOS. This
is because `stpyvista` uses `multiprocessing` under the hood, which has
different defaults. Specifically, its start method on macOS is `spawn()`
rather than `fork()`. This results in issues with PyVista objects
needing to be pickleable.

This PR adds a test with `stpyvista` and a trame backend. `macos-latest`
runners are added to the test matrix. This PR will be ready to merge
when the test passes.

- [x] Try forcing `multiprocessing` start method to use `fork()` always

#### Testing done
<!-- Describe how the changes were tested; e.g., "CI passes", "Tested
manually in stagingrepo#123", screenshots of a terminal session that
verify the changes, or any other evidence of testing the changes. -->
Streamlit's automated testing framework for checking if apps launch
without exceptions.