fix: cast default num vals to their intended types (#26)
#### Description of changes

Added inline type-casting to the default values for `st.number_input`
fields.

This is necessary because Streamlit doesn't allow explicit types to be
set on `number_input`, but instead infers the type based on the given
value. If the intended type is a `float`, but the default is given as an
`int`, Streamlit provides an integer field, and floating points are
unrepresentable.

Casting the defaults to the intended type fixes this issue.

#### Testing done
Manually tested using a Tesseract whose input schema had a parameter
which was typed as a float, but was given as:

- [x] An integer
- [x] A float
- [x] None
- [x] Omitted

Behaves as expected in all cases.

---------

Co-authored-by: Alessandro Angioi <alessandro.angioi@simulation.science>