feat(cli): enforce use of .py ext, or stdout (#25)
#### Description of changes
CLI now throws a `click.UsageError` if `OUTPUT` is either not passed
with a `.py` extension, or is not `stdout` via passing a hyphen `-`.
This has been enforced because Streamlit expects a `.py` extension to
run with the scripts it serves.

#### Testing done
Added an automated test to check that the error is thrown when the
extension is missing, and matches against the specific error message.