refactor(cli): migrate from click to typer (#30)
#### Description of changes

Refactor the CLI from a `click` app to a `typer` app. Migration is just
to make the aesthetic experience of Tesseract Streamlit similar to
Tesseract Core.

> [!NOTE]
> Since `rich` is now used to print the error message introduced in #25,
the control codes make it harder to perform substring matches against
the error message. So, this check has been removed from the automated
test (but the test is still in place to check that an error is thrown).

#### Testing done

Automated tests pass, and manual tests show more consistent visual style
with Tesseract Core.