fix: modify conftest to not rely on Tesseract.serve (#49)
#### Description of changes

Tesseract Core 1.0.0 has a breaking change on the `Tesseract.serve()`
method. It now no longer takes a parameter of `port`.
`tests/conftest.py` relied on this for e2e tests. This has since been
migrated to use the `Tesseract.from_image()`, which does take a `port`
parameter.

#### Testing done
<!-- Describe how the changes were tested; e.g., "CI passes", "Tested
manually in stagingrepo#123", screenshots of a terminal session that
verify the changes, or any other evidence of testing the changes. -->
Run `pytest` locally, and tests now pass with Tesseract Core 1.0.0.