doc: add jax-fem example (#12)
#### Relevant issue or PR
<!-- If the changes resolve an issue or follow some other PR, link to
them here. GitHub references are strongly preferred, but links to other
resources, such as Jira or Confluence, can also be used. Only link
something if it is directly relevant. -->
Closes #4.

#### Description of changes
<!-- Add a high-level description of changes, focusing on the *what* and
*why*. -->
- [X] Move `vectoradd_jax` example to a subdirectory of `examples/`
- [X] Improve README for `vectoradd_jax`
- [X] Add `jax_fem` example, with README, run script, and example JSON
input for bar parameters

Suggested by @dionhaefner 

- [ ] <s>Explore whether we can use default values for arrays to make
`bar_params` auto-populated</s>
- [x] Make the output a spatially resolved field (some FEM solution,
_eg._ stress)
- [x] Add related output plots

#### Testing done
<!-- Describe how the changes were tested; e.g., "CI passes", "Tested
manually in stagingrepo#123", screenshots of a terminal session that
verify the changes, or any other evidence of testing the changes. -->
None yet. Will need to test, since this is an extract of #9, and bits
may be missing.

---------

Co-authored-by: Dion Häfner <mail@dionhaefner.de>
Co-authored-by: Dion Häfner <dion.haefner@simulation.science>
Co-authored-by: Nolan Black <nolan.black@simulation.science>